+++ spark/common/src/main/scala/org/apache/sedona/sql/UDF/Catalog.scala
@@ -327,7 +327,8 @@ object Catalog extends AbstractCatalog with Logging {
     Seq(function[ExpandAddress](), function[ParseAddress]())
 
   // Other / utility expressions not in any docs category
-  val otherExprs: Seq[FunctionDescription] = Seq(function[Barrier]())
+  val otherExprs: Seq[FunctionDescription] =
+    Seq(function[Barrier](), function[__sedona_internal_coverage_invalid_edges_for_target]())
 
   // Aggregate functions implemented as native Catalyst expressions
   val catalystAggregateExprs: Seq[FunctionDescription] = Seq(function[ST_Collect_Agg]())